BDE-226 reviewed tab activation
