+++ src/about/index.js
@@ -18,28 +18,28 @@ limitations under the License.
 import $ from 'jquery';
 import ServiceWorkerBridge from '../service-worker-bridge';
 
-const serverLocator = new ServiceWorkerBridge().serverLocatoor;
+const tabActivator = new ServiceWorkerBridge().serverLocatoor;
 
 $(() => {
   const date = new Date().getFullYear();
   $('#copyright').append(`${date} Nuxeo`);
   $('#apache').click(() => {
-    serverLocator.loadNewExtensionTab(
+    tabActivator.loadNewExtensionTab(
       'http://www.apache.org/licenses/LICENSE-2.0'
     );
   });
   $('#feedback').click(() => {
-    serverLocator.loadNewExtensionTab(
+    tabActivator.loadNewExtensionTab(
       'https://portal.prodpad.com/40c295d6-739d-11e7-9e52-06df22ffaf6f'
     );
   });
   $('#apache').click(() => {
-    serverLocator.loadNewExtensionTab(
+    tabActivator.loadNewExtensionTab(
       'http://www.apache.org/licenses/LICENSE-2.0'
     );
   });
   $('#feedback').click(() => {
-    serverLocator.loadNewExtensionTab(
+    tabActivator.loadNewExtensionTab(
       'https://portal.prodpad.com/40c295d6-739d-11e7-9e52-06df22ffaf6f'
     );
   });
